+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes
 - Fix memleak in UnusedVariableCheck
 - Add MethodDocCheck
 - Extend FormattingCheck to test if line starts with tabs
+- Narrow Magik grammar to be more like SW43
 
 
 0.1.4 (2019-07-26)